+++ tests/evaluateTailwindFunctions.test.js
@@ -1389,3 +1389,30 @@ describe('context dependent', () => {
     })
   })
 })
+
+test('it should handle square brackets inside `theme`, inside arbitrary properties', () => {
+  let config = {
+    content: [
+      {
+        raw: html` <div class="bg-[--color] sm:[--color:_theme(colors.green[400])]"></div> `,
+      },
+    ],
+  }
+
+  let input = css`
+    @tailwind utilities;
+  `
+
+  return runFull(input, config).then((result) => {
+    expect(result.css).toMatchFormattedCss(css`
+      .bg-\[--color\] {
+        background-color: var(--color);
+      }
+      @media (min-width: 640px) {
+        .sm\:\[--color\:_theme\(colors\.green\[400\]\)\] {
+          --color: #4ade80;
+        }
+      }
+    `)
+  })
+})